Rename the filter and test
